Fix incidents and tests

+++ contracts/interfaces/IProductsV1.sol
@@ -3,5 +3,5 @@
 pragma solidity >=0.5.0;
 
 interface IProductsV1 {
-  function getNewProductId(address legacyProductId) external returns (uint24);
+  function getNewProductId(address legacyProductId) external returns (uint);
 }
